Refactor settings page architecture
