--- client/src/components/Settings/components/ModalUIComponent/index.ts
@@ -1,3 +1,0 @@
-import ModalUIComponent from './ModalUIComponent';
-
-export default ModalUIComponent;
