--- client/src/components/Settings/index.ts
@@ -1,19 +1,0 @@
-import Settings from './Settings';
-
-import { connect } from 'react-redux';
-import { createStructuredSelector } from 'reselect';
-
-import { compose } from '../../utils';
-
-import { loading, edgeUsername, user } from '../../redux/profile/selectors';
-
-const reduxConnect = connect(
-  createStructuredSelector({
-    loading,
-    username: edgeUsername,
-    user,
-  }),
-  {},
-);
-
-export default compose(reduxConnect)(Settings);
