+++ client/src/pages/SettingsPage/styles/ModalUIComponent.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/colors.scss';
+@import '../../../assets/styles/colors.scss';
 
 .container {
   max-width: 560px;
