+++ client/src/pages/SettingsPage/styles/PasswordRecovery.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../assets/styles/colors.scss';
-@import '../../../../assets/styles/mediaQueriesMixings.scss';
-@import '../../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/fontMixins.scss';
 
 .resendButton, .changeButton {
   font-size: 0.875em;
