+++ client/src/pages/SettingsPage/styles/SecurityItem.module.scss
@@ -1,6 +1,6 @@
-@import '../../../../assets/styles/fontMixins.scss';
-@import '../../../../assets/styles/colors.scss';
-@import '../../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
 
 .container {
   margin-top: 30px;
