+++ client/src/pages/SettingsPage/styles/SendLinkModal.module.scss
@@ -1,5 +1,5 @@
-@import '../../../../assets/styles/colors.scss';
-@import '../../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
 
 .title {
   margin-top: 30px;
