+++ client/src/pages/SettingsPage/styles/Settings.module.scss
@@ -1,7 +1,7 @@
-@import '../../assets/styles/colors.scss';
-@import '../../assets/styles/fontMixins.scss';
-@import '../../assets/styles/mediaQueriesMixings.scss';
-@import '../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/colors.scss';
+@import '../../../assets/styles/fontMixins.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../assets/styles/mediaQueriesMixings.scss';
 
 .settingsContainer {
   padding: 40px 30px 30px;
